Issue: depending on how tear-down and setup can attempt to delete no longer existing files as it attempts to cleanDb. Current shutdown / setup signature seem to work.
Issue: reseting graph did not cause factories and other to become null. Current
solution: instead of initializing when factory is uninitialized, initialize
when graph database service reference in that class is stale or null.
Desirable: some kind of observer pattern that would be easy to make things
play nicely on service reset:  Low severity, normal running should not be
effected.

Moved Functionality from UniqueEntity extension of Entity into a
UniqueEntityFactory to reduce duplicate code in managing the internal factory
state on Db resets and simplify Unique Entity implementers.

Started Adding Trust Edge, as of this point Email and Citation are
functional but lack navigational methods. User is not yet started.